+++ config/config.exs
@@ -1,6 +1,6 @@
 import Config
 
-config :hermes_mcp, env: :dev
+config :hermes_mcp, env: config_env()
 
 config :logger, :default_formatter,
   format: "[$level] $message $metadata\n",