+++ flake.nix
@@ -1,7 +1,7 @@
 {
   description = "Model Context Porotocol SDK for Elixir";
 
-  inputs.nixpkgs.url = "github:NixOS/nixpkgs/nixos-24.11";
+  inputs.nixpkgs.url = "github:NixOS/nixpkgs/nixos-unstable";
 
   outputs = {nixpkgs, ...}: let
     inherit (nixpkgs.lib) genAttrs;
@@ -15,19 +15,19 @@
       inherit (pkgs.beam) packagesWith;
       beam = packagesWith erlang_27;
       elixir_1_18 = beam.elixir.override {
-        version = "1.18.2";
+        version = "1.18.3";
 
         src = pkgs.fetchFromGitHub {
           owner = "elixir-lang";
           repo = "elixir";
-          rev = "v1.18.2";
-          sha256 = "sha256-8FhUKAaEjBBcF0etVPdkxMfrnR5niU40U8cxDRJdEok=";
+          rev = "v1.18.3";
+          sha256 = "sha256-jH+1+IBWHSTyqakGClkP1Q4O2FWbHx7kd7zn6YGCog0=";
         };
       };
     in {
       default = mkShell {
         name = "hermes-ex";
-        packages = with pkgs; [elixir_1_18 uv just go];
+        packages = with pkgs; [elixir_1_18 uv just go zig _7zz xz];
       };
     });
   };