+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project are documented in this file.
 
+## [0.3.2](https://github.com/cloudwalk/hermes-mcp) - 2025-03-19
+
+### Added
+- Ship static binaries to use hermes-mcp as standlone application
+
+## [0.3.1](https://github.com/cloudwalk/hermes-mcp) - 2025-03-19
+
+### Added
+- Ship interactive mix tasks `stdio.interactive` and `sse.interactive` to test MCP servers
+
 ## [0.3.0](https://github.com/cloudwalk/hermes-mcp) - 2025-03-18
 
 ### Added